+++ CHANGELOG.md
@@ -70,6 +70,7 @@ high state of flux, you're at risk of it changing without notice.
     - make `member` pipeable (@gcanti)
     - make `elem` pipeable (@gcanti)
     - make `lookupWithKey` pipeable (@gcanti)
+    - make `lookup` pipeable (@gcanti)
   - `ReadonlyNonEmptyArray`
     - make `cons` pipeable (@gcanti)
     - make `snoc` pipeable (@gcanti)